feat : implement Start interview, save response to redis and database and endpoint for FE

+++ src/main/java/fpt/org/inblue/repository/InterviewSessionRepository.java
@@ -3,5 +3,5 @@ package fpt.org.inblue.repository;
 import fpt.org.inblue.model.InterviewSession;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface InterviewSessionRepository extends JpaRepository<InterviewSession, Long> {
+public interface InterviewSessionRepository extends JpaRepository<InterviewSession, Integer> {
 }
